Split duplicate pillow tests into their own test file
